debug phase 1

+++ matlab/slam/robot/motion/v2pL.m
@@ -1,3 +1,3 @@
 function pL = v2pL(v)
-    pL = round(v * (-122.00) - 16.75);
+    pL = int32(round(v * (-122.00) - 16.75));
 end
